refactor(crypto): split crypto/secrets/storage.rs into per-section submodules
Convert `crypto/secrets/storage.rs` (1193 LOC) into `storage/` with three
per-section submodules:

- `basic.rs` (346 LOC): `SecretStorage` struct + sync impl + `Drop` +
  `Debug` + `Default`
- `managed.rs` (406 LOC): `ManagedStorageConfig`, `ManagedSecretStorage`
  (async wrapper with background tokio cleanup task), and
  `ManagedStorageBuilder` + the `DEFAULT_CLEANUP_INTERVAL` constant
- `tests.rs` (394 LOC): all `#[cfg(test)]` unit tests

`storage/mod.rs` (75 LOC) re-exports `SecretStorage`, `ManagedSecretStorage`,
`ManagedStorageBuilder`, `ManagedStorageConfig` so the
`crypto::secrets::*` API surface is unchanged.

`SecretStorage.{secrets, storage_id}` and `ManagedSecretStorage.config`
were bumped from private to `pub(super)` to support cross-submodule
access; semantically identical visibility (same crate, same `storage/`
module).

Pure move-only refactor — no logic, signature, or external API changes.
Follows the precedent set by #317/#318/#331/#335/#341/#342/#343/#344.

Part of #338 (file-length umbrella). 5 of 7 PRs.
